Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
pillow
opencv-python
tensorflow